Updated GUI, added file upload.
Added webpages for About, Contact, and Source.  Also added an upload
feature on the main webpage, which will take .txt, .png, .jpg, .gif, and
upload it to the uploads/apk_samples/ folder.  Currently if the file
already exists it will be overwritten, so... we'll need to add
functionality to detect duplicate filenames.

Also updated the README to contain 'TODO' for simple milestones
we need to get done