Added 'mock' reporting functionality
malDroid will now accept .APK files and insert some meta-data about the
file into the database.  The schema.sql file has been updated to
accurately reflect all, in my opinion - I could be wrong, of the data
which we'll need to insert for a given sample.  The code also now has
a point of entry for where we can begin the malware analysis.  Also
added a few new templates and some malware samples to be used for
testing.